backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiomysql
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiohttp
httpx
google-generativeai
psutil
